clean up sglang and vllm integrations
