+++ lib/gpu_memory_service/integrations/__init__.py
@@ -1,2 +1,4 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
+
+"""GPU Memory Service integrations for inference frameworks."""
