+++ lib/gpu_memory_service/integrations/vllm/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""GPU Memory Service integration for vLLM.
+
+Usage:
+    Set --load-format gms --worker-cls gpu_memory_service.integrations.vllm.worker:GMSWorker
+"""
